requirements.txt
msal>=1.20
requests>=2.28
asana>=1.0
pydantic>=1.10.0
httpx>=0.24.0
tenacity>=8.2.2
structlog>=23.1.0
transformers>=4.30.0
pymdptoolbox>=4.11
python-dotenv>=1.0
beautifulsoup4>=4.11.1
colorama>=0.4.6
pandas>=1.5
matplotlib>=3.5
scikit-learn>=1.2
nltk>=3.8
textblob>=0.17
pdfminer.six>=20240524
python-docx>=0.8.11
openpyxl>=3.1.2
pillow>=11.0.0
pytesseract>=0.3.10
